pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram_recovery_tool"
version = "0.1.0"
authors = [
    { name="André de Freitas Smaira", email="andre.smaira@gmail.com" },
]
description = "This tool recovers deleted messages of a group if you are admin and if they were deleted less then 48h ago and generates a file with both the deleted and not deleted messages, ordered in time order."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "python-dotenv",
    "telethon",
    "json",
    "datetime",
    "asyncio"
]

[project.urls]
Homepage = "https://github.com/afsmaira/telegram-recovery-tool"

[project.scripts]
telegramrec = "telegram_recovery_tool.main:main"